+++ admob/app/src/main/AndroidManifest.xml
@@ -19,7 +19,9 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+        <activity android:name=".BannerActivity"/>
         <activity android:name=".InterstitialActivity"/>
+        <activity android:name=".RewardActivity"/>
     </application>
 
 </manifest>
